+++ desktop-widgets/tab-widgets/TabDiveStatistics.ui
@@ -197,7 +197,7 @@
   <customwidget>
    <class>MinMaxAvgWidget</class>
    <extends>QWidget</extends>
-   <header>desktop-widgets/simplewidgets.h</header>
+   <header>desktop-widgets/tab-widgets/TabDiveStatistics.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
